Move types to their own source files
